pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uds3"
version = "1.4.0"
description = "Unified Database Strategy v3 - Enterprise Multi-Database Distribution System"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Proprietary"}
authors = [
    {name = "UDS3 Development Team"}
]
keywords = ["database", "multi-db", "vector-search", "graph-database", "rag", "search-api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # chromadb>=0.4.0 - REMOVED: We use remote HTTP API only (via requests)
    "requests>=2.31.0",  # For ChromaDB Remote HTTP Client
    "neo4j>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/uds3"
Documentation = "https://github.com/yourusername/uds3/blob/main/README.md"
Repository = "https://github.com/yourusername/uds3"
Changelog = "https://github.com/yourusername/uds3/blob/main/CHANGELOG.md"

[tool.setuptools]
py-modules = [
    "uds3_core",
    "uds3_search_api",
    "uds3_dsgvo_core",
    "uds3_security_quality",
    "uds3_saga_orchestrator",
    "uds3_streaming_operations",
    "uds3_polyglot_query",
    "uds3_identity_service",
    "uds3_naming_strategy",
    "uds3_multi_db_distributor",
    "adaptive_multi_db_strategy",
    "config"
]
packages = ["search", "database", "tools"]

[tool.setuptools.package-data]
search = ["*.py"]
database = ["*.py"]
tools = ["*.py"]

[tool.ruff]
line-length = 88
exclude = [
  "**/__pycache__/**",
  "**/.venv/**",
  "docs/**",
  "third_party_stubs/**",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["D"]

